Inline headers from demo-interactor
